Rename otcollector to opencensus
