+++ ext/opentelemetry-ext-opencensus/MANIFEST.in
