+++ ext/opentelemetry-ext-opencensus/README.rst
@@ -3,8 +3,8 @@ OpenTelemetry Collector Exporter
 
 |pypi|
 
-.. |pypi| image:: https://badge.fury.io/py/opentelemetry-ext-otcollector.svg
-   :target: https://pypi.org/project/opentelemetry-ext-otcollector/
+.. |pypi| image:: https://badge.fury.io/py/opentelemetry-ext-opencensus.svg
+   :target: https://pypi.org/project/opentelemetry-ext-opencensus/
 
 This library allows to export data to `OpenTelemetry Collector`_ , currently using OpenCensus receiver in Collector side.
 
@@ -13,12 +13,12 @@ Installation
 
 ::
 
-     pip install opentelemetry-ext-otcollector
+     pip install opentelemetry-ext-opencensus
 
 
 References
 ----------
 
-* `OpenTelemetry Collector Exporter <https://opentelemetry-python.readthedocs.io/en/latest/ext/otcollector/otcollector.html>`_
+* `OpenTelemetry Collector Exporter <https://opentelemetry-python.readthedocs.io/en/latest/ext/opencensus/opencensus.html>`_
 * `OpenTelemetry Collector <https://github.com/open-telemetry/opentelemetry-collector/>`_
 * `OpenTelemetry <https://opentelemetry.io/>`_
