+++ ext/opentelemetry-ext-opencensus/setup.cfg
@@ -13,13 +13,13 @@
 # limitations under the License.
 #
 [metadata]
-name = opentelemetry-ext-otcollector
+name = opentelemetry-ext-opencensus
 description = OpenTelemetry Collector Exporter
 long_description = file: README.rst
 long_description_content_type = text/x-rst
 author = OpenTelemetry Authors
 author_email = cncf-opentelemetry-contributors@lists.cncf.io
-url = https://github.com/open-telemetry/opentelemetry-python/tree/master/ext/opentelemetry-ext-otcollector
+url = https://github.com/open-telemetry/opentelemetry-python/tree/master/ext/opentelemetry-ext-opencensus
 platforms = any
 license = Apache-2.0
 classifiers =
