+++ ext/opentelemetry-ext-opencensus/setup.py
@@ -17,7 +17,7 @@ import setuptools
 
 BASE_DIR = os.path.dirname(__file__)
 VERSION_FILENAME = os.path.join(
-    BASE_DIR, "src", "opentelemetry", "ext", "otcollector", "version.py"
+    BASE_DIR, "src", "opentelemetry", "ext", "opencensus", "version.py"
 )
 PACKAGE_INFO = {}
 with open(VERSION_FILENAME) as f:
