+++ ext/opentelemetry-ext-opencensus/tests/test_otcollector_metrics_exporter.py
@@ -20,7 +20,7 @@ from google.protobuf.timestamp_pb2 import Timestamp
 from opencensus.proto.metrics.v1 import metrics_pb2
 
 from opentelemetry import metrics
-from opentelemetry.ext.otcollector import metrics_exporter
+from opentelemetry.ext.opencensus import metrics_exporter
 from opentelemetry.sdk.metrics import (
     Counter,
     Measure,
@@ -47,7 +47,7 @@ class TestCollectorMetricsExporter(unittest.TestCase):
     def test_constructor(self):
         mock_get_node = mock.Mock()
         patch = mock.patch(
-            "opentelemetry.ext.otcollector.util.get_node",
+            "opentelemetry.ext.opencensus.util.get_node",
             side_effect=mock_get_node,
         )
         service_name = "testServiceName"
