+++ tox.ini
@@ -59,9 +59,9 @@ envlist =
     py3{4,5,6,7,8}-test-ext-mysql
     pypy3-test-ext-mysql
 
-    ; opentelemetry-ext-otcollector
-    py3{4,5,6,7,8}-test-ext-otcollector
-    ; ext-otcollector intentionally excluded from pypy3
+    ; opentelemetry-ext-opencensus
+    py3{4,5,6,7,8}-test-ext-opencensus
+    ; ext-opencensus intentionally excluded from pypy3
 
     ; opentelemetry-ext-prometheus
     py3{4,5,6,7,8}-test-ext-prometheus
@@ -143,7 +143,7 @@ changedir =
   test-ext-dbapi: ext/opentelemetry-ext-dbapi/tests
   test-ext-django: ext/opentelemetry-ext-django/tests
   test-ext-mysql: ext/opentelemetry-ext-mysql/tests
-  test-ext-otcollector: ext/opentelemetry-ext-otcollector/tests
+  test-ext-opencensus: ext/opentelemetry-ext-opencensus/tests
   test-ext-prometheus: ext/opentelemetry-ext-prometheus/tests
   test-ext-pymongo: ext/opentelemetry-ext-pymongo/tests
   test-ext-psycopg2: ext/opentelemetry-ext-psycopg2/tests
@@ -184,7 +184,7 @@ commands_pre =
 
   mysql: pip install {toxinidir}/opentelemetry-auto-instrumentation {toxinidir}/ext/opentelemetry-ext-dbapi {toxinidir}/ext/opentelemetry-ext-mysql[test]
 
-  otcollector: pip install {toxinidir}/ext/opentelemetry-ext-otcollector
+  opencensus: pip install {toxinidir}/ext/opentelemetry-ext-opencensus
 
   prometheus: pip install {toxinidir}/ext/opentelemetry-ext-prometheus
